fix: refactor daily settlement aggregation to use Redis Lua scripts and optimize item processing
