+++ batch_service/src/main/java/com/project/batch_service/jobs/daily_settle/dto/SettlementAggregationResult.java
@@ -1,0 +1,29 @@
+package com.project.batch_service.jobs.daily_settle.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.math.BigDecimal;
+import java.time.LocalDate;
+
+@Data
+@AllArgsConstructor
+public class SettlementAggregationResult {
+
+    private Long settlementId;
+    private Long sellerId;
+    private LocalDate settlementDate;
+    private Integer totalOrderCount;
+    private Integer totalClaimCount;
+    private Integer totalQuantity;
+    private BigDecimal totalSalesAmount;
+    private BigDecimal totalTaxAmount;
+    private BigDecimal totalPromotionDiscountAmount;
+    private BigDecimal totalCouponDiscountAmount;
+    private BigDecimal totalPointUsedAmount;
+    private BigDecimal totalShippingFee;
+    private BigDecimal totalClaimShippingFee;
+    private BigDecimal totalCommissionAmount;
+    private BigDecimal totalSettlementAmount;
+
+}
